[fix] registerするとDB上のuser.bonitosがintになるバグを修正
DB上のbonitosはdoubleの前提で他の部分がコーディングされているため
intになると他のAPIがコケてしまう

+++ src/jp/kobe_u/cspiral/norakore/model/User.java
@@ -47,7 +47,7 @@ public class User {
         if (bo == null) throw new Exception("user's bonitos doesn't exist on DB.");
         Double bonitos_d = (Double)bo;
         int bonitos = bonitos_d.intValue() + new_bonitos;
-        user.put("bonitos", bonitos);
+        user.put("bonitos", (double)bonitos);
         return bonitos;
     }
